pyproject.toml
[project]
name = "hirokihamaguchi-github-io"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "getorg>=0.3.1",
    "ipyleaflet>=0.20.0",
    "ipywidgets>=8.1.7",
    "numpy>=2.3.3",
    "pandas>=2.3.3",
    "pillow>=12.1.0",
    "requests>=2.32.4",
    "types-requests>=2.32.4.20250913",
]
